# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Binar Challenge 05 - Cars",
    "description": "this is simple crud car api that implemented authentication and return data from PosgreSQL database.\n\nrepo:\n- [binar challenge repo](https://github.com/hasyimashari/binar-challenge.git)",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://127.0.0.1:3000"
    }
  ],
...
  "paths": {
...
    "/api/cars": {
      "get": {
        "tags": [
          "cars"
        ],
        "summary": "get cars",
        "description": "get list of car",
        "operationId": "getListCars",
        "responses": {
          "200": {
            "description": "operation success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "OK"
                    },
                    "message": {
                      "type": "string",
                      "example": "get user data success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/List_cars_return"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "operation failed - unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/API_respon"
                },
                "example": {
                  "status": "FAIL",
                  "message": "Failed to access: unauthorized"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "cars"
        ],
        "summary": "create car",
        "description": "create new car item",
        "operationId": "createCar",
        "requestBody": {
          "description": "data to create new car",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Car_input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "operation success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Car_return"
                }
              }
            }
          },
          "400": {
            "description": "operation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/API_respon"
                },
                "examples": {
                  "null data": {
                    "value": {
                      "status": "FAIL",
                      "message": "Failed to create new user : ${data} is required"
                    }
                  },
                  "empty data": {
                    "value": {
                      "status": "FAIL",
                      "message": "Failed to create new user : ${data} cant be empty"
                    }
                  },
                  "wrong car type": {
                    "value": {
                      "status": "FAIL",
                      "message": "Failed to created new car : car type must be either small, medium, or large"
                    }
                  },
                  "not number data": {
                    "value": {
                      "status": "FAIL",
                      "message": "Failed to created new car : ${data} can be only numbers"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "operation failed - unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/API_respon"
                },
                "example": {
                  "status": "FAIL",
                  "message": "Failed to access: unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "operation failed - forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/API_respon"
                },
                "example": {
                  "status": "FAIL",
                  "message": "Failed to access: forbidden"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Car_input": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Ford F150"
          },
          "type": {
            "type": "string",
            "enum": [
              "small",
              "medium",
              "large"
            ],
            "example": "small"
          },
          "image": {
            "type": "string",
            "format": "url",
            "example": "./images/car01.min.jpg"
          },
          "capacity": {
            "type": "number",
            "example": 2
          },
          "rentPerDay": {
            "type": "number",
            "example": 20000
          },
          "description": {
            "type": "string",
            "example": "Brake assist. Leather-wrapped shift knob. Glove box lamp. Air conditioning w/in-cabin microfilter."
          },
          "availableAt": {
            "type": "string",
            "format": "date-time",
            "example": "2022-03-23T15:49:05.563Z"
          }
        },
        "required": [
          "name",
          "type",
          "image",
          "capacity",
          "rentPerDay",
          "availableAt"
        ]
      },
      "Car_return": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "5dbbf4b9-ff96-477a-a542-e33c866d6332\""
          },
          "name": {
            "type": "string",
            "example": "Ford F150"
          },
          "type": {
            "type": "string",
            "enum": [
              "small",
              "medium",
              "large"
            ],
            "example": "small"
          },
          "image": {
            "type": "string",
            "format": "url",
            "example": "./images/car01.min.jpg"
          },
          "capacity": {
            "type": "number",
            "example": 2
          },
          "rentPerDay": {
            "type": "number",
            "example": 20000
          },
          "description": {
            "type": "string",
            "example": "Brake assist. Leather-wrapped shift knob. Glove box lamp. Air conditioning w/in-cabin microfilter."
          },
          "availableAt": {
            "type": "string",
            "format": "date-time",
            "example": "2022-03-23T15:49:05.563Z"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2022-03-23T15:49:05.563Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": "2022-03-23T15:49:05.563Z"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": "2022-03-23T15:49:05.563Z"
          },
          "createdBy": {
            "type": "string",
            "format": "uuid",
            "example": "5dbbf4b9-ff96-477a-a542-e33c866d6332\""
          },
          "updatedBy": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "example": "5dbbf4b9-ff96-477a-a542-e33c866d6332\""
          },
          "deletedBy": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "example": "5dbbf4b9-ff96-477a-a542-e33c866d6332\""
          }
        }
      },
      "List_cars_return": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "format": "uuid",
              "example": "5dbbf4b9-ff96-477a-a542-e33c866d6332\""
            },
            "name": {
              "type": "string",
              "example": "Ford F150"
            },
            "type": {
              "type": "string",
              "enum": [
                "small",
                "medium",
                "large"
              ],
              "example": "small"
            },
            "image": {
              "type": "string",
              "format": "url",
              "example": "./images/car01.min.jpg"
            },
            "capacity": {
              "type": "number",
              "example": 2
            },
            "rentPerDay": {
              "type": "number",
              "example": 20000
            },
            "description": {
              "type": "string",
              "example": "Brake assist. Leather-wrapped shift knob. Glove box lamp. Air conditioning w/in-cabin microfilter."
            },
            "availableAt": {
              "type": "string",
              "format": "date-time",
              "example": "2022-03-23T15:49:05.563Z"
            }
          }
        }
      },
...
      "API_respon": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "OK",
              "FAIL"
            ]
          },
          "message": {
            "type": "string"
          },
          "data": {
            "nullable": true
          }
        }
      }
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}